Finished refactoring. First pass at implementing the /ui/amputationLevels endpoint.
