+++ .meteor/packages
@@ -4,17 +4,17 @@
 # 'meteor add' and 'meteor remove' will edit this file for you,
 # but you can also edit it by hand.
 
-meteor-base@1.0.4             # Packages every Meteor app needs to have
+meteor-base@1.1.0             # Packages every Meteor app needs to have
 mobile-experience@1.0.4       # Packages for a great mobile UX
-mongo@1.1.16                   # The database Meteor supports right now
+mongo@1.1.18                   # The database Meteor supports right now
 blaze-html-templates@1.0.4    # Compile .html files into Meteor Blaze views
 jquery@1.11.10                  # Helpful client-side library
-tracker@1.1.2                 # Meteor's client-side reactive programming library
+tracker@1.1.3                 # Meteor's client-side reactive programming library
 
 standard-minifier-css@1.3.4   # CSS minifier run for production mode
-standard-minifier-js@1.2.3    # JS minifier run for production mode
+standard-minifier-js@2.1.0    # JS minifier run for production mode
 es5-shim@4.6.15                # ECMAScript 5 compatibility for older browsers.
-ecmascript@0.6.3              # Enable ECMAScript2015+ syntax in app code
+ecmascript@0.8.0              # Enable ECMAScript2015+ syntax in app code
 shell-server@0.2.3            # Server-side component of the `meteor shell` command
 
 npm-bcrypt@0.9.2
@@ -30,9 +30,10 @@ u2622:persistent-session
 session@1.1.7
 fortawesome:fontawesome@=4.7.0
 accounts-ui@1.1.9
-accounts-password@1.3.4
+accounts-password@1.3.6
 materialize:materialize
 perak:codemirror
 gantrim:meteor-json-view-materialize
 okgrow:analytics
 planettraining:material-design-icons
+dynamic-import
