+++ .meteor/versions
@@ -1,34 +1,37 @@
-accounts-base@1.2.15
-accounts-password@1.3.4
+accounts-base@1.3.0
+accounts-password@1.3.6
 accounts-ui@1.1.9
 accounts-ui-unstyled@1.2.1
 allow-deny@1.0.5
 amplify@1.0.0
 autoupdate@1.3.12
-babel-compiler@6.14.1
+babel-compiler@6.19.2
 babel-runtime@1.0.1
 base64@1.0.10
 binary-heap@1.0.10
-blaze@2.3.1
-blaze-html-templates@1.1.1
+blaze@2.3.2
+blaze-html-templates@1.1.2
 blaze-tools@1.0.10
-boilerplate-generator@1.0.11
+boilerplate-generator@1.1.0
 caching-compiler@1.1.9
-caching-html-compiler@1.1.1
+caching-html-compiler@1.1.2
 callback-hook@1.0.10
 check@1.2.5
 coffeescript@1.12.3_1
 ddp@1.2.5
-ddp-client@1.3.3
+ddp-client@1.3.4
 ddp-common@1.2.8
 ddp-rate-limiter@1.0.7
-ddp-server@1.3.13
+ddp-server@1.3.14
 deps@1.0.12
 diff-sequence@1.0.7
-ecmascript@0.6.3
-ecmascript-runtime@0.3.15
+dynamic-import@0.1.1
+ecmascript@0.8.0
+ecmascript-runtime@0.4.1
+ecmascript-runtime-client@0.4.1
+ecmascript-runtime-server@0.4.1
 ejson@1.0.13
-email@1.1.18
+email@1.2.3
 es5-shim@4.6.15
 fastclick@1.0.13
 fortawesome:fontawesome@4.7.0
@@ -40,20 +43,20 @@ html-tools@1.0.11
 htmljs@1.0.11
 http@1.2.12
 id-map@1.0.9
-jagi:astronomy@2.4.1
+jagi:astronomy@2.4.4
 jquery@1.11.10
 kadira:blaze-layout@2.3.0
 kadira:flow-router@2.12.1
 launch-screen@1.1.1
 less@2.7.9
 livedata@1.0.18
-localstorage@1.0.12
+localstorage@1.1.0
 logging@1.1.17
 maodouio:logs@0.0.4
-materialize:materialize@0.98.0
+materialize:materialize@0.98.2
 mdg:validation-error@0.5.1
 meteor@1.6.1
-meteor-base@1.0.4
+meteor-base@1.1.0
 meteortoys:allthings@3.0.0
 meteortoys:authenticate@3.0.0
 meteortoys:autopub@3.0.0
@@ -70,13 +73,13 @@ meteortoys:sub@3.0.0
 meteortoys:throttle@3.0.0
 meteortoys:toykit@3.0.4
 minifier-css@1.2.16
-minifier-js@1.2.18
-minimongo@1.0.21
+minifier-js@2.1.0
+minimongo@1.2.0
 mobile-experience@1.0.4
 mobile-status-bar@1.0.14
-modules@0.7.9
-modules-runtime@0.7.9
-mongo@1.1.16
+modules@0.9.1
+modules-runtime@0.8.0
+mongo@1.1.18
 mongo-id@1.0.6
 msavin:jetsetter@2.0.0
 msavin:mongol@2.0.1
@@ -90,10 +93,10 @@ ordered-dict@1.0.9
 perak:codemirror@1.3.1
 planettraining:material-design-icons@2.2.3
 pntbr:js-yaml-client@0.0.1
-promise@0.8.8
+promise@0.8.9
 random@1.0.10
-rate-limit@1.0.7
-reactive-dict@1.1.8
+rate-limit@1.0.8
+reactive-dict@1.1.9
 reactive-var@1.0.11
 reload@1.1.11
 retry@1.0.9
@@ -103,19 +106,19 @@ session@1.1.7
 sha@1.0.9
 shell-server@0.2.3
 simple:json-routes@2.1.0
-spacebars@1.0.14
-spacebars-compiler@1.1.1
+spacebars@1.0.15
+spacebars-compiler@1.1.2
 srp@1.0.10
 standard-minifier-css@1.3.4
-standard-minifier-js@1.2.3
-templating@1.3.1
-templating-compiler@1.3.1
-templating-runtime@1.3.1
-templating-tools@1.1.1
-tracker@1.1.2
+standard-minifier-js@2.1.0
+templating@1.3.2
+templating-compiler@1.3.2
+templating-runtime@1.3.2
+templating-tools@1.1.2
+tracker@1.1.3
 u2622:persistent-session@0.4.4
-ui@1.0.12
+ui@1.0.13
 underscore@1.0.10
 url@1.1.0
-webapp@1.3.14
+webapp@1.3.16
 webapp-hashing@1.0.9
