+++ client/index.html
@@ -1,5 +1,5 @@
 <head>
-  <title>Fusion Microservice</title>
+  <title>Shift Microservice</title>
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
 </head>
