+++ imports/routes/server.js
@@ -1,0 +1,4 @@
+import './api/healthcheck.js';
+import './api/preview.js';
+import './api/rex.js';
+import './api/ui.js';
