+++ imports/startup/client.js
@@ -1,3 +1,3 @@
 import './both.js';
 
-// import '/imports/api/mixpanel.js'
+import '/imports/api/mixpanel.js'
