+++ package.json
@@ -7,7 +7,7 @@
   },
   "dependencies": {
     "agenda": "^0.9.0",
-    "babel-runtime": "^6.20.0",
+    "babel-runtime": "^6.23.0",
     "bcrypt": "^0.8.7",
     "fibers": "^1.0.15",
     "hasha": "^2.2.0",
@@ -25,4 +25,3 @@
     "typedarray-to-buffer": "^3.1.2"
   }
 }
-
